Fixed store in Todos, loading TodoList
