+++ src/app/state/app.state.ts
@@ -1,0 +1,3 @@
+export interface State {
+    loggedId: boolean;
+}
