+++ src/app/todo/state/todo.selectors.ts
@@ -1,0 +1,10 @@
+import { TodoState } from './todo.state';
+import { createFeatureSelector, createSelector } from '@ngrx/store';
+
+
+const getTodoFeatureState = createFeatureSelector<TodoState>('todos');
+
+export const getTodoList = createSelector(
+    getTodoFeatureState,
+    state => state.todos
+);
