+++ src/app/todo/state/todo.state.ts
@@ -1,0 +1,17 @@
+import { Todo } from '../models/todo.model';
+import * as fromRoot from '../../state/app.state';
+
+
+export interface State extends fromRoot.State {
+    todos: TodoState;
+}
+
+export interface TodoState {
+    todos: Todo[];
+    error: string;
+}
+
+export const initialState: TodoState = {
+    todos: [],
+    error: ''
+};
