+++ src/app/todo/todo-list/todo-list.component.ts
@@ -2,7 +2,9 @@ import { OnInit, Component } from '@angular/core';
 import { Todo } from '../models/todo.model';
 import { Store,  select} from '@ngrx/store';
 import * as todoActions from '../state/todo.actions';
-import { TodoState, State } from '../state/todo.reducer';
+import { Observable } from 'rxjs';
+import { State } from '../state/todo.state';
+import { getTodoList } from '../state/todo.selectors';
 
 @Component({
     selector: 'kros-todo-list',
@@ -12,15 +14,15 @@ import { TodoState, State } from '../state/todo.reducer';
 export class TodoListComponent implements OnInit {
 
     constructor(
-        private store: Store<State>) { }
+        private store: Store<State>
+    ) {
+    }
 
-    todoList: Todo[] = [];
+    todoList$: Observable<Todo[]>;
 
     ngOnInit() {
         this.store.dispatch(new todoActions.Load());
-        this.store.pipe(select(state => state.todos)).subscribe(
-            (x: TodoState) => {
-                this.todoList = x.todos;
-            });
+
+        this.todoList$ = this.store.pipe(select(getTodoList));
     }
 }
